feat(add-seventh): Redid the logic for adding seventh + changed the progression bank to suit it.
We now do not consider Dom7, Dim7 or Min7b5 as a chord that requires the whole prog to have a 7th.
We now only add a seventh if base had a seventh, not considering sub.
Changed the progression bank, removed some 7th that would be added if the user wrote the prog with 7th.